Improve user experience by changing the fragments handling
Without this, the fragments could only be a list of files,
and was merged into a list through a temporary variable.

This intends to put the user defined fragments at the forefront
of the support, and make sure it is very flexible in the features,
while not making the code too complex.