Add recommended keywords
haproxy config check will by default work if there is a listener.

This is a problem as it might not be the default behaviour a user
wants to see: If for example the stats listener is working, the
frontends/backends/listens stanzas of the user might not work.

This brings a list of the _minimum_ expectations of keywords.
A user will be able to ignore recommendations.