Add generation of the global configuration section
Without this, globals are not generated. Now we have a test,
a validation of user input, and a working global system with
merges at the right levels.

This also required renaming the test scenario from
"no-extra-fragments" to "cleanup-fragments" for clarity.

At the same time, the tests in the future would fail
due to the absence of recommended information
(the timeouts/binds), so this was updated too.