Ensure the facts are clean at the start of the run
When haproxy_others_fragments are set through default/vars,
its content gets overriden in the loop with the set_fact.

This is however a problem, when multiple plays are running, as
set_facts persist between plays. So it means that multiple
plays passing different variables to the same host could have
data from a different run.

This fixes it by ensuring every run has an empty set of facts
at the start.