Change notify/handler behaviour
Without this patch, any change in config would trigger a "notify"
of restart haproxy.

This is a problem, because a change in the local config could be
invalid (for the sake of testing) and trigger a restart of haproxy
(with an already assembled configuration).

This fixes it by ensuring the reconfiguration of haproxy only
happens if the rest of the role ran fine (implemented by the
addition of a conditional in handlers) and by ensuring that a
restart only happens when the files are assembled together.