Add docker support in tests
This will allow to run tests everywhere.
All the tests that worked on localhost now needed to be moved
to run in a remote container, hence the addition of
an inventory, and the docker collection.

Some tests that relied on localhost behaviour will
continue to work on localhost, and now need to be delegated to
localhost.

Some tests relied on behaviour of failure, checking the failure
facts. Sadly, while run_once is supposed to set facts for all
the nodes it replaces, it seems that ansible_failed_task
is not defined in all nodes, and the tests have started to
break. To workaround this ansible bug, I will simply run the
plays on the first node. This run_once will always be the node
with the facts.